+++ src/charts/AreaChart.vue
@@ -106,7 +106,6 @@ const config = computed<AxisChartConfig>(() => ({
   ...normalized.value.config,
   type: 'area',
   stacked: props.stacked,
-  connectNulls: props.connectNulls,
 }))
 
 const {